High quality screen shots in Unity.
Using this script you can take high quality screen shots in unity during game play. In the script first you have to select the key for which the screen would be captured, and you will have to specify the location of the pictures. The script allows you to take snap shots during game play without pausing it or anything like that. Just attach it to your main canvas and specify the key and address and enjoy :)
PS: Do make sure to remove the script after taking the screen shots and before creating the build.