Display FPS on the screen.
This script displays the FPS(Frames Per Second) of your game/app. Simply add the script to any of the active game object and you are good to go. You can also change the size of the FPS display from script. This script is the best for testing FPS of your games.